Adds setMultilineTokens to ContiguousTokensStore
