+++ src/vs/editor/common/tokens/contiguousTokensStore.ts
@@ -10,6 +10,8 @@ import { ContiguousTokensEditing, EMPTY_LINE_TOKENS, toUint32Array } from 'vs/ed
 import { LineTokens } from 'vs/editor/common/tokens/lineTokens';
 import { ILanguageIdCodec } from 'vs/editor/common/languages';
 import { LanguageId, FontStyle, ColorId, StandardTokenType, MetadataConsts, TokenMetadata } from 'vs/editor/common/encodedTokenAttributes';
+import { ITextModel } from 'vs/editor/common/model';
+import { ContiguousMultilineTokens } from 'vs/editor/common/tokens/contiguousMultilineTokens';
 
 /**
  * Represents contiguous tokens in a text model.
@@ -207,6 +209,39 @@ export class ContiguousTokensStore {
 	}
 
 	//#endregion
+
+	public setMultilineTokens(tokens: ContiguousMultilineTokens[], textModel: ITextModel): { changes: { fromLineNumber: number; toLineNumber: number }[] } {
+		if (tokens.length === 0) {
+			return { changes: [] };
+		}
+
+		const ranges: { fromLineNumber: number; toLineNumber: number }[] = [];
+
+		for (let i = 0, len = tokens.length; i < len; i++) {
+			const element = tokens[i];
+			let minChangedLineNumber = 0;
+			let maxChangedLineNumber = 0;
+			let hasChange = false;
+			for (let lineNumber = element.startLineNumber; lineNumber <= element.endLineNumber; lineNumber++) {
+				if (hasChange) {
+					this.setTokens(textModel.getLanguageId(), lineNumber - 1, textModel.getLineLength(lineNumber), element.getLineTokens(lineNumber), false);
+					maxChangedLineNumber = lineNumber;
+				} else {
+					const lineHasChange = this.setTokens(textModel.getLanguageId(), lineNumber - 1, textModel.getLineLength(lineNumber), element.getLineTokens(lineNumber), true);
+					if (lineHasChange) {
+						hasChange = true;
+						minChangedLineNumber = lineNumber;
+						maxChangedLineNumber = lineNumber;
+					}
+				}
+			}
+			if (hasChange) {
+				ranges.push({ fromLineNumber: minChangedLineNumber, toLineNumber: maxChangedLineNumber, });
+			}
+		}
+
+		return { changes: ranges };
+	}
 }
 
 function getDefaultMetadata(topLevelLanguageId: LanguageId): number {
